Add the possibility to use a bag file instead of a physical device
A bag_filename can now be specified as parameter to the
orbbec_camera_node, that will be replayed using the PlaybackDevice
interface.

A bag_replay_loop boolean parameter is added to allow replaying the bag
in a loop.

The format of the bag file is the one generated when recording from the
OrbbecViewer tool.